requirements.txt
transformers
torchvision 
torchaudio
streamlit
python-docx
environs
--index-url https://download.pytorch.org/whl/cu126